Pure cookie session store, multiple simultaneous logins
For simple out of the box usage, add a cookie store for the required state.
This is unencrypted, so the refresh and access tokens are not tracked - just
the ID token. The state and PKCE tokens are stored here too, there should be no
forgery risks here but apps should be aware of this. A proper store is still
reccomended for most cases.

Track multiple in-process logins - this is a fairly common case, and stops the
invalid state errors.