Update claims for Tink JWT
It has better methods for getting and setting values by the expected type, and
is generally stricter and more explicit. Move to adoping this, to avoid
weirdness in our old setup (e.g is groups in extra a []any, or a []string?). We
still provide convenient structures for the standard claims, but these are now
directly built from / marshaled to a tink RawJWT.