Re-work token cache
As with everything else, make it align a bit more with x/oauth2. Make the cache
key more generic so it can be used with oauth2 as well, add a helper to build a
key for OIDC. Move the current cache implementations to clitoken, they're more
CLI aligned - but the cache can be used with server stuff too.