Better keychain handling
Keychain sets an ACL to only allow the creating executable to read entries.
When it's a different executable, the user has to enter their password to
permit access. This is good, but it's a UX nightmare when the binary changes,
because there can be many secrets that need re-auth.

Fortunately, we're using this more as a short-term cache, rather than wanting a
secret to live for months. Hopefully obtaining a fresh one isn't too hard
either. So lean on the side of clean UX and only read entries created by this
instance of the executable, and delete the entry if it was created by a
different one. Because different binaries might use the same issuer/client,
also prefix the entry with the name of the binary that created it. This will
allow them to co-exist without too much interruption.

At some point in the future we might change this for long-lived refresh tokens
or something by making it opt out, for now this is a good balance.

Update the tests to validate this more thoroughly too.