Don't bypass keyset fetch in verification
The verifier is trying to shortcut re-fetching keys by testing the current
cached keyset, but then fails to fall back to fetching if that verification
fails. This can get processes in a state where they're perpetually using stale
data. Remove this path, and allow the normal re-fetch to proceed as needed. We
get rid of the shortcut altogether, as this could also potentially result in
the process validating with revoked keys.